Rename implementation files for Class* to Classic*
