+++ src/ClassOption/ClassicOption.d.ts
